Extract UIFactory from UIManager
