+++ src/ts/demofactory.ts
@@ -32,6 +32,7 @@ import {SubtitleListBox} from './components/subtitlelistbox';
 import {AudioTrackListBox} from './components/audiotracklistbox';
 import {SettingsPanelItem} from './components/settingspanelitem';
 import {SettingsPanelPage} from './components/settingspanelpage';
+import { UIFactory } from './uifactory';
 
 export namespace DemoFactory {
 
@@ -40,18 +41,18 @@ export namespace DemoFactory {
     let smallScreenSwitchWidth = 600;
 
     return new UIManager(player, [{
-      ui: UIManager.Factory.modernSmallScreenAdsUI(),
+      ui: UIFactory.modernSmallScreenAdsUI(),
       condition: (context: UIConditionContext) => {
         return context.isMobile && context.documentWidth < smallScreenSwitchWidth
           && context.isAd && context.adClientType === 'vast';
       },
     }, {
-      ui: UIManager.Factory.modernAdsUI(),
+      ui: UIFactory.modernAdsUI(),
       condition: (context: UIConditionContext) => {
         return context.isAd && context.adClientType === 'vast';
       },
     }, {
-      ui: UIManager.Factory.modernSmallScreenUI(),
+      ui: UIFactory.modernSmallScreenUI(),
       condition: (context: UIConditionContext) => {
         return context.isMobile && context.documentWidth < smallScreenSwitchWidth;
       },
